requirements.txt
moviepy
openai-whisper
playwright
Pillow
numpy 